api/requirements.txt
asgiref==3.6.0
attrs==22.2.0
dj-database-url==0.5.0
Django==3.2.18
django-cors-headers==3.5.0
django-on-heroku==1.1.2
djangorestframework==3.12.2
djangorestframework-simplejwt==5.2.2
drf-spectacular==0.25.1
gunicorn==20.0.4
importlib-resources==5.12.0
inflection==0.5.1
install==1.3.5
jsonschema==4.17.3
pkgutil_resolve_name==1.3.10
psycopg2-binary==2.9.5
PyJWT==2.6.0
pyrsistent==0.19.3
python-dotenv==0.21.1
pytz==2022.7.1
PyYAML==6.0
sqlparse==0.4.3
uritemplate==4.1.1
whitenoise==5.2.0
zipp==3.14.0
